Restructure database files add linter
