+++ .eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = { extends: ['airbnb-base', 'prettier'] };
