+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules/
-dummyData.js
+.vscode
+database/dummyData.js
