+++ database/connection.js
@@ -1,6 +1,9 @@
 const mongoose = require('mongoose');
 
 const mongoUri = 'mongodb://localhost/fec-descriptions';
-const db = mongoose.connect(mongoUri, { useNewUrlParser: true });
+const db = mongoose.connect(mongoUri, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+});
 
 module.exports = db;
