+++ package.json
@@ -10,10 +10,12 @@
   "scripts": {
     "server": "nodemon ./server/app.js --ignore node_modules",
     "react-dev": "webpack -d --watch",
-    "seed": "node database/fakerGenerate.js"
+    "seed": "node database/seed.js"
   },
   "dependencies": {
+    "eslint": "^7.2.0",
     "express": "^4.17.1",
+    "faker": "^4.1.0",
     "mongoose": "^5.9.19",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
@@ -23,11 +25,9 @@
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.10.1",
     "babel-loader": "^8.1.0",
-    "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "eslint-plugin-import": "^2.21.2",
-    "faker": "^4.1.0",
     "nodemon": "^2.0.4"
   }
 }
