+++ server/app.js
@@ -1,4 +1,5 @@
 const express = require('express');
+
 const app = express();
 
 app.use(express.static('public'));
@@ -8,5 +9,5 @@ app.get('/', (req, res) => {
 });
 
 app.listen(3000, () => {
-  console.log('Listening');
+  console.log('Listening on port 3000!');
 });
